requirements.txt
fastapi
uvicorn
python-multipart
requests
PyPDF2
Pillow
pytesseract
python-dotenv
google-generativeai